Constants have been created for drawable and screens.json Strings are taken out of the code in strings.xml end:Int renamed to gamePassed:Boolean Transition.kt renamed to ScenesLoad.kt Packages ui and model have been created
