+++ app/src/main/AndroidManifest.xml
@@ -10,13 +10,13 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.VisualNovel">
         <activity
-            android:name=".GameActivity"
+            android:name=".ui.GameActivity"
             android:exported="false" />
         <activity
-            android:name=".NamingActivity"
+            android:name=".ui.NamingActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.MainActivity"
             android:exported="true"
             android:screenOrientation="portrait">
             <intent-filter>
