+++ app/src/main/java/com/example/visualnovel/ui/GameActivity.kt
@@ -1,30 +1,22 @@
-package com.example.visualnovel
+package com.example.visualnovel.ui
 
 import android.content.Intent
-import android.opengl.Visibility
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.util.Log
 import android.view.View
-import android.widget.Toast
+import com.example.visualnovel.model.ScenesLoad
 import com.example.visualnovel.databinding.GameMainBinding
-import kotlinx.serialization.decodeFromString
-import kotlinx.serialization.json.*
-import kotlinx.serialization.*
-import kotlinx.serialization.encodeToString
-import kotlinx.serialization.json.Json
-import kotlinx.serialization.Serializable
 
 class GameActivity : AppCompatActivity() {
 
-    private lateinit var transition:Transition
+    private lateinit var scenesLoad: ScenesLoad
     private lateinit var binding:GameMainBinding
     private lateinit var name:String
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = GameMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
-        transition=Transition(this)
+        scenesLoad= ScenesLoad(this)
         name=intent.getStringExtra("username").toString()
         nextScene(0)
         binding.buttonVariant1.setOnClickListener{nextScene(0)}
@@ -37,16 +29,16 @@ class GameActivity : AppCompatActivity() {
         binding.buttonVariant2.visibility= View.INVISIBLE
         binding.buttonVariant3.visibility= View.INVISIBLE
 
-        val data=transition.playerChoice(next)
+        val data=scenesLoad.playerChoice(next)
 
         if(data.id==0){
             val intent = Intent(this, MainActivity::class.java)
-            intent.putExtra("end",1)
+            intent.putExtra("gameOver",true)
             startActivity(intent)
         }
 
         binding.textView.text=data.message.replace("<username>",name)
-        binding.backgroundImage.setImageResource(resources.getIdentifier(data.image,"drawable",packageName))
+        binding.backgroundImage.setImageResource(resources.getIdentifier(data.image, DRAWABLE_DEFTYPE,packageName))
         binding.buttonVariant1.text=data.choices[0].text
         binding.buttonVariant1.visibility= View.VISIBLE
 
@@ -58,5 +50,9 @@ class GameActivity : AppCompatActivity() {
             binding.buttonVariant3.text=data.choices[2].text
             binding.buttonVariant3.visibility= View.VISIBLE
         }
+    }
+
+    companion object {
+        private const val DRAWABLE_DEFTYPE="drawable"
     }
 }
