+++ app/src/main/java/com/example/visualnovel/ui/NamingActivity.kt
@@ -1,9 +1,10 @@
-package com.example.visualnovel
+package com.example.visualnovel.ui
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Toast
+import com.example.visualnovel.R
 import com.example.visualnovel.databinding.NamingMainBinding
 
 class NamingActivity : AppCompatActivity() {
@@ -13,7 +14,7 @@ class NamingActivity : AppCompatActivity() {
         setContentView(binding.root)
         binding.confirmButton.setOnClickListener{
             if(binding.usernameTextView.text.toString()==""){
-                Toast.makeText(this,"Пожалуйста, введи имя!",Toast.LENGTH_LONG).show()
+                Toast.makeText(this,getString(R.string.empty_name),Toast.LENGTH_LONG).show()//
             }else{
                 val intent = Intent(this, GameActivity::class.java)
                 intent.putExtra("username",binding.usernameTextView.text.toString())
