+++ app/src/main/res/layout/game_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#000000"
-    tools:context=".GameActivity">
+    tools:context=".ui.GameActivity">
 
     <ImageView
         android:id="@+id/backgroundImage"
