+++ app/src/main/res/layout/naming_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#000000"
-    tools:context=".NamingActivity">
+    tools:context=".ui.NamingActivity">
 
     <ImageView
         android:id="@+id/backgroundImage"
